rafthttp: always cancel in-flight request when pipeline.send
This fits the way for go1.5 to cancel request.
